use u64 for timestamp

+++ bee-api/bee-rest-api/src/types/responses.rs
@@ -138,8 +138,7 @@ pub struct OutputResponse {
     #[serde(rename = "milestoneIndex")]
     pub milestone_index: u32,
     #[serde(rename = "milestoneTimestamp")]
-    // todo: u32 or u64 here? in MilestoneResponse it's an u64
-    pub milestone_timestamp: u32,
+    pub milestone_timestamp: u64,
     #[serde(rename = "ledgerIndex", default)]
     pub ledger_index: u32,
     pub output: OutputDto,
